GEODE-9950: Add LRANGE command
This implements a version of the Redis LRANGE command, which is used for list data types. Associated tests were also added.

Given a key, it returns a list of elements in the range (inclusive) for the specified start index and stop index. Both the start and stop indexes are zero based, which starts at the head of the list. Negative indexes start at the tail of the list.

LINDEX was modified to use a new method created in this commit. The method converts negative indexes to the corresponding positive index.